+++ CHANGELOG.md
@@ -5,9 +5,11 @@
 * Update messages written to stdout by the `delete-my-tweets` and
  `delete-my-likes` commands.
 
-* Add `--no-links` option to the `archive-timeline` command.
+* Add `--[no|only]-retweets` options to the `archive-timeline` command.
 
-* Add `--no-media` option to the `archive-timeline` command.
+* Add `--[no|only]-links` options to the `archive-timeline` command.
+
+* Add `--[no|only]-media` options to the `archive-timeline` command.
 
 * Add `--outbound-links-only` option to the `archive-timeline` command.